face_cut/requirements.txt
ydb
pillow